Setup codegen struct
